set tests into different suites to run 1 at a time
