+++ package.json
@@ -12,7 +12,7 @@
   "scripts": {
     "start": "react-scripts start",
     "server": "json-server --watch db.json -p 6001",
-    "test": "jest"
+    "test": "jest --bail"
   },
   "eslintConfig": {
     "extends": [
