+++ src/__tests__/SearchPlants.test.js
@@ -1,0 +1,25 @@
+import React from 'react';
+import { render, fireEvent, waitFor } from '@testing-library/react';
+import App from '../components/App';
+import '@testing-library/jest-dom';
+
+describe('Plantsy app', () => {
+    test('filters plants by name on search', async () => {
+        global.setFetchResponse(global.basePlants)
+        const { getByPlaceholderText, queryAllByTestId } = render(<App />);
+        const searchInput = getByPlaceholderText('Type a name to search...');
+        fireEvent.change(searchInput, { target: { value: 'aloe' } });
+    
+        await waitFor(() => {
+          const filteredPlants = queryAllByTestId('plant-item');
+          expect(filteredPlants).toHaveLength(1);
+        });
+        
+        fireEvent.change(searchInput, { target: { value: 'p' } });
+        
+        await waitFor(() => {
+          const filteredPlants = queryAllByTestId('plant-item');
+          expect(filteredPlants).toHaveLength(3);
+        });
+      });
+})
